move win32 low-level file io funcs to fopen.c (from multibyte.c)
