+++ lib/curlx/fopen.h
@@ -26,6 +26,8 @@
 
 #include "../curl_setup.h"
 
+#include "multibyte.h"
+
 #if defined(_WIN32) && !defined(UNDER_CE)
 FILE *curlx_win32_fopen(const char *filename, const char *mode);
 #define CURLX_FOPEN_LOW(fname, mode) curlx_win32_fopen(fname, mode)
